Add basic integration test on KinD cluster
This is a low functionality proof of concept. It uses GitHub actions to do the following:

 1. Create a KinD k8s cluster in the Ubuntu box
 2. Runs kubewise with go run main.go
 3. Installs and uninstalls the zookeeper helm package.
 4. Saves the logs from KubeWise.

The idea is that the messages can be viewed in Slack and asserted for
correctness.

This isn't a nice solution and needs a lot more work. I need to learn
more about Go testing to figure out the best way to assert the output.