Send initialization messages and events
When KubeWise first starts up, there is no indication to the user that
it is working. They don't know if it worked until they install a Helm
chart. This may not happen for quite some time and they may not want to
install or uninstall something just to see a message. This commit adds a
welcome message which is sent when the app starts up. It will also be
sent if the pod is restarted.

 - Initialization message will be sent for both Slack and Google Chat
 - An initialization webhook request will be sent.
 - The code is now using the default Helm drivers. `decodeRelease` is
   no-longer copy/pasted out of the Helm codebase.
 - One of the presenter files is renamed from `presenter.go` to
   `text.go`.
 - Check docker is running before building